core/images/file-storage-service/src/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.12.14
cffi==1.17.1
grpcio==1.70.0
grpcio-tools==1.70.0
minio==7.2.15
protobuf==5.29.3
pycparser==2.22
pycryptodome==3.21.0
setuptools==75.8.0
typing_extensions==4.12.2
urllib3==2.3.0
prometheus_client